# Make devices::Naive::matmul_fw_impl() faster.