About 2 times faster than the older one.
Benchmark on my machine:
- CPU: Core i7-6700K 4GHz
- Threads: 1
- Code:
```c++
#include <iostream>
#include <primitiv/primitiv.h>
using namespace std;
using namespace primitiv;
namespace F = primitiv::operators;
main() {
  devices::Naive dev(0);
  Device::set_default(dev);
  const Tensor x = F::random::uniform<Tensor>({1024, 1024}, -1, 1);
  const Tensor y = F::random::uniform<Tensor>({1024, 1024}, -1, 1);
  for (int i = 1; i <= 100; ++i) {
    cout << i << '\r' << flush;
    volatile const Tensor z = F::matmul(x, y);
  }
  cout << "done" << endl;
}
```
- Result:
```
before: 154.33s user 0.01s system 99% cpu 2:34.37 total
after:   73.55s user 0.00s system 99% cpu 1:13.57 total
```