@odashi 2017-11-24
Related to: #68